pseudo modularizing bessel diff
