+++ include/eve/module/bessel/diff/impl/airy_bi.hpp
@@ -7,9 +7,11 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/function/airy_bi.hpp>
-#include <eve/function/derivative.hpp>
-#include <eve/function/oneminus.hpp>
+
+#include <eve/module/math.hpp>
+#include <eve/module/bessel/regular/cyl_bessel_jn.hpp>
+#include <eve/module/bessel/regular/cyl_bessel_in.hpp>
+#include <eve/detail/hz_device.hpp>
 
 namespace eve::detail
 {
