+++ include/eve/module/bessel/diff/impl/cyl_bessel_i1.hpp
@@ -7,10 +7,9 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/function/cyl_bessel_in.hpp>
-#include <eve/function/cyl_bessel_i0.hpp>
-#include <eve/function/average.hpp>
-#include <eve/function/derivative.hpp>
+#include <eve/module/math.hpp>
+#include <eve/module/bessel/regular/cyl_bessel_in.hpp>
+#include <eve/module/bessel/regular/cyl_bessel_i0.hpp>
 
 namespace eve::detail
 {
