+++ include/eve/module/bessel/diff/impl/cyl_bessel_k0.hpp
@@ -7,8 +7,7 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/function/cyl_bessel_k1.hpp>
-#include <eve/function/derivative.hpp>
+#include <eve/module/bessel/regular/cyl_bessel_k1.hpp>
 
 namespace eve::detail
 {
