+++ include/eve/module/bessel/diff/impl/cyl_bessel_k1.hpp
@@ -7,9 +7,9 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/function/cyl_bessel_kn.hpp>
-#include <eve/function/cyl_bessel_k0.hpp>
-#include <eve/function/derivative.hpp>
+#include <eve/module/math.hpp>
+#include <eve/module/bessel/regular/cyl_bessel_k0.hpp>
+#include <eve/module/bessel/regular/cyl_bessel_kn.hpp>
 
 namespace eve::detail
 {
