+++ include/eve/module/bessel/diff/impl/cyl_bessel_y0.hpp
@@ -7,8 +7,8 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/function/cyl_bessel_y1.hpp>
-#include <eve/function/derivative.hpp>
+#include <eve/module/math.hpp>
+#include <eve/module/bessel/regular/cyl_bessel_y1.hpp>
 
 namespace eve::detail
 {
