+++ include/eve/module/bessel/diff/impl/cyl_bessel_y1.hpp
@@ -7,9 +7,9 @@
 //==================================================================================================
 #pragma once
 
-#include <eve/function/cyl_bessel_yn.hpp>
-#include <eve/function/cyl_bessel_y0.hpp>
-#include <eve/function/derivative.hpp>
+#include <eve/module/math.hpp>
+#include <eve/module/bessel/regular/cyl_bessel_yn.hpp>
+#include <eve/module/bessel/regular/cyl_bessel_y0.hpp>
 
 namespace eve::detail
 {
